Dry up decryptEmbeddedDocs
