+++ lib/plugins/encrypted-children.js
@@ -1,13 +1,8 @@
 'use strict';
 
 var mongoose = require('mongoose');
-var dotty = require('dotty');
-var _ = require('underscore');
 
-var objectUtil = require('../util/object-util.js');
-var cryptoUtil = require('../util/crypto-util.js');
-
-var isEmbeddedDocument = objectUtil.isEmbeddedDocument;
+var decryptEmbeddedDocs = require('../util/decryptEmbeddedDocs');
 
 /**
  * Export For Schemas That Contain Encrypted Embedded Documents
@@ -19,21 +14,6 @@ var isEmbeddedDocument = objectUtil.isEmbeddedDocument;
 
 module.exports = function(schema) {
     schema.post('save', function(doc) {
-    	// TODO DRY
-    	_.keys(schema.paths).forEach(function (path) {
-    	    if (path === '_id' || path === '__v') {
-    	        return;
-    	    }
-
-    	    var nestedDoc = dotty.get(doc, path);
-
-    	    if (nestedDoc && nestedDoc[0] && isEmbeddedDocument(nestedDoc[0])) {
-    	        nestedDoc.forEach(function (subDoc) {
-    	            if (_.isFunction(subDoc.decryptSync)) {
-    	                subDoc.decryptSync();
-    	            }
-    	        });
-    	    }
-    	});
+    	doc.decryptEmbeddedDocs();
     });
 };
