+++ lib/plugins/mongoose-encryption.js
@@ -12,6 +12,7 @@ var semver = require('semver');
 
 var objectUtil = require('../util/object-util.js');
 var cryptoUtil = require('../util/crypto-util.js');
+var decryptEmbeddedDocs = require('../util/decryptEmbeddedDocs.js');
 
 var pick = objectUtil.pick;
 var setFieldValue = objectUtil.setFieldValue;
@@ -301,7 +302,7 @@ var mongooseEncryption = function(schema, options) {
                     doc.decryptSync();
                 }
 
-                doc._decryptEmbeddedDocs();
+                decryptEmbeddedDocs(doc);
 
                 return doc;
             });
@@ -389,31 +390,6 @@ var mongooseEncryption = function(schema, options) {
             this._ct = undefined;
             this._ac = undefined;
         }
-    };
-
-    /**
-     * Decrypt any embedded documents inside of this document
-     *
-     * @private
-     * Should not be needed outside plugin
-     */
-    schema.methods._decryptEmbeddedDocs = function () {
-        var doc = this;
-        _.keys(this.schema.paths).forEach(function (path) {
-            if (path === '_id' || path === '__v') {
-                return;
-            }
-
-            var nestedDoc = dotty.get(doc, path);
-
-            if (nestedDoc && nestedDoc[0] && isEmbeddedDocument(nestedDoc[0])) {
-                nestedDoc.forEach(function (subDoc) {
-                    if (_.isFunction(subDoc.decryptSync)) {
-                        subDoc.decryptSync();
-                    }
-                });
-            }
-        });
     };
